Fixed Bugs
