+++ bot/plugins/authorize.py
@@ -15,9 +15,11 @@ from bot.helpers.utils import CustomFilters
 
 OAUTH_SCOPE = "https://www.googleapis.com/auth/drive"
 REDIRECT_URI = "urn:ietf:wg:oauth:2.0:oob"
-G_DRIVE_DIR_MIME_TYPE = "application/vnd.google-apps.folder"
+
+#This credentials are taken from https://rclone.org change it accordingly.
 G_DRIVE_CLIENT_ID = "202264815644.apps.googleusercontent.com"
 G_DRIVE_CLIENT_SECRET = "X4Z3ca8xfWDb1Voo-F9a7ZxJ"
+
 flow = None
 
 @Client.on_message(filters.private & filters.incoming & filters.command(BotCommands.Authorize))
@@ -54,7 +56,7 @@ def _revoke(client, message):
   user_id = message.from_user.id
   try:
     gDriveDB._clear(user_id)
-    LOGGER.info(f'Revoke:{user_id}')
+    LOGGER.info(f'Revoked:{user_id}')
     message.reply_text(Messages.REVOKED, quote=True)
   except Exception as e:
     message.reply_text(f"**ERROR:** ```{e}```", quote=True)
